+++ x/bucket/keeper/msg_server_create_bucket.go
@@ -2,6 +2,8 @@ package keeper
 
 import (
 	"context"
+	"fmt"
+	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -38,14 +40,18 @@ func (k msgServer) CreateBucket(goCtx context.Context, msg *types.MsgCreateBucke
 
 	// Create a new channel record
 	newWhichIs := types.WhichIs{
-		Did:     did.ID,
-		Creator: msg.GetSession().Creator(),
-		Bucket:  doc,
+		Did:       did.ID,
+		Creator:   msg.GetSession().Creator(),
+		Bucket:    doc,
+		Timestamp: time.Now().Unix(),
+		IsActive:  true,
 	}
 
 	// Store the channel record
 	k.SetWhichIs(ctx, newWhichIs)
 	return &types.MsgCreateBucketResponse{
 		WhichIs: &newWhichIs,
+		Code:    100,
+		Message: fmt.Sprintf("New Bucket %s has been created", newWhichIs.Did),
 	}, nil
 }