+++ x/object/keeper/msg_server_create_object.go
@@ -2,6 +2,8 @@ package keeper
 
 import (
 	"context"
+	"fmt"
+	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -47,11 +49,15 @@ func (k msgServer) CreateObject(goCtx context.Context, msg *types.MsgCreateObjec
 		Did:       did.ID,
 		Creator:   msg.GetSession().Creator(),
 		ObjectDoc: doc,
+		Timestamp: time.Now().Unix(),
+		IsActive:  true,
 	}
 
 	// Store the Object record
 	k.SetWhatIs(ctx, newWhatIs)
 	return &types.MsgCreateObjectResponse{
-		WhatIs: &newWhatIs,
+		WhatIs:  &newWhatIs,
+		Code:    100,
+		Message: fmt.Sprintf("New Object %s has been created", newWhatIs.Did),
 	}, nil
 }