+++ x/object/types/message_deactivate_object.go
@@ -5,26 +5,26 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-const TypeMsgDeleteObject = "delete_object"
+const TypeMsgDeactivateObject = "delete_object"
 
-var _ sdk.Msg = &MsgDeleteObject{}
+var _ sdk.Msg = &MsgDeactivateObject{}
 
-func NewMsgDeleteObject(creator string, did string, publicKey string) *MsgDeleteObject {
-	return &MsgDeleteObject{
+func NewMsgDeactivateObject(creator string, did string, publicKey string) *MsgDeactivateObject {
+	return &MsgDeactivateObject{
 		Creator: creator,
 		Did:     did,
 	}
 }
 
-func (msg *MsgDeleteObject) Route() string {
+func (msg *MsgDeactivateObject) Route() string {
 	return RouterKey
 }
 
-func (msg *MsgDeleteObject) Type() string {
-	return TypeMsgDeleteObject
+func (msg *MsgDeactivateObject) Type() string {
+	return TypeMsgDeactivateObject
 }
 
-func (msg *MsgDeleteObject) GetSigners() []sdk.AccAddress {
+func (msg *MsgDeactivateObject) GetSigners() []sdk.AccAddress {
 	creator, err := sdk.AccAddressFromBech32(msg.Creator)
 	if err != nil {
 		panic(err)
@@ -32,12 +32,12 @@ func (msg *MsgDeleteObject) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{creator}
 }
 
-func (msg *MsgDeleteObject) GetSignBytes() []byte {
+func (msg *MsgDeactivateObject) GetSignBytes() []byte {
 	bz := ModuleCdc.MustMarshalJSON(msg)
 	return sdk.MustSortJSON(bz)
 }
 
-func (msg *MsgDeleteObject) ValidateBasic() error {
+func (msg *MsgDeactivateObject) ValidateBasic() error {
 	_, err := sdk.AccAddressFromBech32(msg.Creator)
 	if err != nil {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)